+++ footnotes/mod.ts
@@ -257,7 +257,7 @@ export default function footNotes(md: any, userOptions: Partial<Options> = {}) {
         currentFootnote!.content = md.renderer.render(
           currentTokens,
           state.md.options,
-          state.env
+          state.env,
         );
         currentTokens = undefined;
         return false;