+++ utils.ts
@@ -17,3 +17,9 @@ export function getRawText(tokens: any[]) {
 
   return text;
 }
+
+export function slugify(x: string): string {
+  return encodeURIComponent(
+    String(x).trim().toLowerCase().replace(/\s+/g, "-"),
+  );
+}