+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.9.0] - 2025-03-04
+### Changed
+- Simplified Wikilinks plugin.
+
+### Fixed
+- Pass `state.env` in `footnote_reference_close` blocks [#2]
+
 ## [0.8.0] - 2024-12-31
 ### Added
 - Wikilinks plugin.
@@ -65,6 +72,9 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - `toc` plugin.
 - `title` plugin.
 
+[#2]: https://github.com/lumeland/markdown-plugins/issues/2
+
+[0.9.0]: https://github.com/lumeland/markdown-plugins/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/lumeland/markdown-plugins/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/lumeland/markdown-plugins/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/lumeland/markdown-plugins/compare/v0.6.0...v0.7.0